Moved JSON files from `src/json` to .js exports in `src/data` (issue #166)
