+++ src/data/cookies_usercustom.js
@@ -1,0 +1,17 @@
+/*
+OptMeowt is licensed under the MIT License
+Copyright (c) 2020 Kuba Alicki, Daniel Knopf, Abdallah Salia, Sebastian Zimmeck
+privacy-tech-lab, https://privacytechlab.org/
+*/
+
+/*
+cookies_usercustom.js
+================================================================================
+cookies_usercustom.js holds all cookies that are chosen by the user. Please follow
+the format of the default cookies stored in cookies_3p.js
+*/
+
+
+export const cookies_usercustom = {
+
+}
