--- src/json/cookies_usercustom.json
@@ -1,3 +1,0 @@
-{
-    
-}
