--- src/json/headers.json
@@ -1,10 +1,0 @@
-{
-   "Sec-GPC": {
-      "name": "Sec-GPC",
-      "value": "1"
-   },
-   "DNT": {
-      "name": "DNT",
-      "value": "1"
-   }
-}
