+++ src/options/options.html
@@ -20,15 +20,15 @@ options.html is the html scaffolding for OptMeowt's options page
     <title>OptMeowt</title>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="../libs-css/uikit.min.css" />
-    <link rel="stylesheet" href="../libs-css/animate.compat.css" />
-    <link rel="stylesheet" href="styles.css" />
-    <script src="../libs-js/uikit.min.js"></script>
-    <script src="../libs-js/uikit-icons.min.js"></script>
-    <script src="../libs-js/mustache.js"></script>
-    <script src="../libs-js/popper.js"></script>
-    <script src="../libs-js/tippy-bundle.umd.min.js"></script>
-    <script type="module" src="options.js"></script>
+    <!-- <link rel="stylesheet" href="../libs-css/uikit.min.css" /> -->
+    <!-- <link rel="stylesheet" href="../libs-css/animate.compat.css" /> -->
+    <!-- <link rel="stylesheet" href="styles.css" /> -->
+    <!-- <script src="../libs-js/uikit.min.js"></script> -->
+    <!-- <script src="../libs-js/uikit-icons.min.js"></script> -->
+    <!-- <script src="../libs-js/mustache.js"></script> -->
+    <!-- <script src="../libs-js/popper.js"></script> -->
+    <!-- <script src="../libs-js/tippy-bundle.umd.min.js"></script> -->
+    <!-- <script type="module" src="options.js"></script> -->
   </head>
 
   <!-- This is the entry point to 'Options' -->
